backend/requirements.txt
asgiref==3.5.2
black==22.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
